+++ packages/ai/src/telemetry/telemetry-options.ts
